// src/gadgets/editing/lyrics-editor/hottable.d.ts
type PredefinedMenuItemKey = 'row_above' | 'row_below' | 'col_left' | 'col_right' |
  '---------' | 'remove_row' | 'remove_col' | 'clear_column' | 'undo' | 'redo' |
  'make_read_only' | 'alignment' | 'cut' | 'copy' | 'copy_column_headers_only' |
  'copy_with_column_group_headers' | 'copy_with_column_headers' | 'freeze_column' |
  'unfreeze_column' | 'borders' | 'commentsAddEdit' | 'commentsRemove' | 'commentsReadOnly' |
  'mergeCells' | 'add_child' | 'detach_from_parent' | 'hidden_columns_hide' | 'hidden_columns_show' |
  'hidden_rows_hide' | 'hidden_rows_show' | 'filter_by_condition' | 'filter_operators' |
  'filter_by_condition2' | 'filter_by_value' | 'filter_action_bar';
type ContextMenuSettings = boolean | PredefinedMenuItemKey[] | DetailedSettings;
export interface CellSelection {
  start: CellCoords;
  end: CellCoords;
}
export type ContextMenuCallback = (key: string, selection: CellSelection[], clickEvent: MouseEvent) => void;
interface DetailedSettings {
  callback?: ContextMenuCallback,
  uiContainer?: HTMLElement,
  items?: PredefinedMenuItemKey[] | MenuConfig;
}
// ...
export interface CellCoords {
  constructor(row: number, column: number, isRtl?: boolean);

  row: number;
  col: number;

  isValid(tableParams: {
    countRows?: number;
    countCols?: number;
    countRowHeaders?: number;
    countColHeaders?: number;
  }): boolean;
  isRtl(): boolean;
  isEqual(cellCoords: CellCoords): boolean;
  isSouthEastOf(testedCoords: CellCoords): boolean;
  isNorthWestOf(testedCoords: CellCoords): boolean;
  isSouthWestOf(testedCoords: CellCoords): boolean;
  isNorthEastOf(testedCoords: CellCoords): boolean;
  normalize(): CellCoords;
  assign(coords: CellCoords | { row?: number, col?: number }): CellCoords;
  clone(): CellCoords;
  toObject(): { row: number, col: number };
}